feat: restructure proj and enhance ui
