+++ web/index.html
@@ -1,0 +1,209 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Meme SSE Debugger</title>
+    <style>
+        :root {
+            --bg-primary: #f4f4f4;
+            --bg-secondary: #ffffff;
+            --text-primary: #333;
+            --border-color: #e0e0e0;
+            --connection-online: #4CAF50;
+            --connection-offline: #F44336;
+        }
+
+        * {
+            box-sizing: border-box;
+            margin: 0;
+            padding: 0;
+        }
+
+        body {
+            font-family: 'Arial', sans-serif;
+            line-height: 1.6;
+            background-color: var(--bg-primary);
+            color: var(--text-primary);
+            display: flex;
+            flex-direction: column;
+            min-height: 100vh;
+            padding: 1rem;
+        }
+
+        .container {
+            display: flex;
+            flex-direction: column;
+            gap: 1rem;
+            max-width: 1200px;
+            margin: 0 auto;
+            width: 100%;
+        }
+
+        .meme-container {
+            background-color: var(--bg-secondary);
+            border-radius: 8px;
+            padding: 1rem;
+            text-align: center;
+            box-shadow: 0 4px 6px rgba(0,0,0,0.1);
+        }
+
+        .meme-image {
+            max-width: 100%;
+            max-height: 70vh;
+            object-fit: contain;
+            border-radius: 4px;
+            margin: 1rem 0;
+        }
+
+        .connection-status {
+            display: inline-block;
+            padding: 0.5rem 1rem;
+            border-radius: 4px;
+            font-weight: bold;
+        }
+
+        .connection-online {
+            background-color: var(--connection-online);
+            color: white;
+        }
+
+        .connection-offline {
+            background-color: var(--connection-offline);
+            color: white;
+        }
+
+        .debug-container {
+            background-color: var(--bg-secondary);
+            border-radius: 8px;
+            padding: 1rem;
+            box-shadow: 0 4px 6px rgba(0,0,0,0.1);
+        }
+
+        .debug-log {
+            background-color: var(--bg-primary);
+            border-radius: 4px;
+            padding: 1rem;
+            margin-bottom: 1rem;
+            word-break: break-all;
+        }
+
+        .debug-log-header {
+            font-weight: bold;
+            margin-bottom: 0.5rem;
+            color: var(--text-primary);
+        }
+
+        .debug-log-detail {
+            display: flex;
+            margin-bottom: 0.25rem;
+        }
+
+        .debug-log-label {
+            font-weight: bold;
+            margin-right: 0.5rem;
+            min-width: 120px;
+        }
+
+        @media (min-width: 768px) {
+            body {
+                flex-direction: row;
+            }
+
+            .container {
+                flex-direction: row;
+            }
+
+            .meme-container {
+                flex: 2;
+                margin-right: 1rem;
+            }
+
+            .debug-container {
+                flex: 1;
+                max-height: 100vh;
+                overflow-y: auto;
+            }
+        }
+    </style>
+</head>
+<body>
+    <div class="container">
+        <div class="meme-container">
+            <h1>Meme SSE Debugger</h1>
+            <div id="connectionStatus" class="connection-status connection-offline">
+                Disconnected
+            </div>
+            <h2 id="memeTitle">Waiting for memes...</h2>
+            <img id="memeImage" class="meme-image" src="" alt="Meme">
+            <p id="connectionID">Connection ID: N/A</p>
+        </div>
+        <div class="debug-container">
+            <h2>Debug Logs</h2>
+            <div id="debugLogs"></div>
+        </div>
+    </div>
+
+    <script>
+        const connectionStatusEl = document.getElementById('connectionStatus');
+        const eventSource = new EventSource('/memes');
+        const titleEl = document.getElementById('memeTitle');
+        const imageEl = document.getElementById('memeImage');
+        const connectionIDEl = document.getElementById('connectionID');
+        const debugLogsEl = document.getElementById('debugLogs');
+
+        function updateConnectionStatus(isConnected) {
+            connectionStatusEl.textContent = isConnected ? 'Connected' : 'Disconnected';
+            connectionStatusEl.className = `connection-status ${isConnected ? 'connection-online' : 'connection-offline'}`;
+        }
+
+        function fetchDebugLogs() {
+            fetch('/debug')
+                .then(response => response.json())
+                .then(logs => {
+                    debugLogsEl.innerHTML = logs.map(log => `
+                        <div class="debug-log">
+                            <div class="debug-log-header">Connection ${log.id}</div>
+                            <div class="debug-log-detail">
+                                <span class="debug-log-label">Remote:</span>
+                                <span>${log.remote_addr}</span>
+                            </div>
+                            <div class="debug-log-detail">
+                                <span class="debug-log-label">Timestamp:</span>
+                                <span>${new Date(log.timestamp).toLocaleString()}</span>
+                            </div>
+                            <div class="debug-log-detail">
+                                <span class="debug-log-label">Events:</span>
+                                <span>${log.events.join(' → ')}</span>
+                            </div>
+                        </div>
+                    `).join('');
+                })
+                .catch(error => {
+                    console.error('Failed to fetch debug logs:', error);
+                });
+        }
+
+        eventSource.onopen = function() {
+            updateConnectionStatus(true);
+        };
+
+        eventSource.onmessage = function(event) {
+            const meme = JSON.parse(event.data);
+            titleEl.textContent = meme.title;
+            imageEl.src = meme.url;
+            connectionIDEl.textContent = 'Connection ID: ' + meme.connID;
+            fetchDebugLogs();
+        };
+
+        eventSource.onerror = function(error) {
+            console.error('EventSource failed:', error);
+            updateConnectionStatus(false);
+            eventSource.close();
+        };
+
+        // Initial log fetch
+        fetchDebugLogs();
+    </script>
+</body>
+</html>
